Ensure that a socket does not get used after close
If one writer thread encounters an error while sending to Cassandra,
causing the connection to get torn down and recreated, it is possible
for another writer thread using the same connection to attempt to
write to the socket after the reader thread closes it.  At best this
causes an EBADF error; at worst it could send data to a newly opened
connection or file that happens to use the same FD.

This uses the write-lock that already exists to both guard the
close operation and carry a flag that the writers can inspect to
ensure the socket is still open before doing its write.

"ResourceVanished" is probably a better error than IllegalOperation,
though that's a GHC-specific IO error.  It won't get propagated to user
code anyway; it'll just get caught in the IOException catch-all and
cause a retry.