`inUse` vs. resource sequence length.
`take1` may observe that the `inUse` counter is different to the
resource sequence length, an invariant that should generally hold.
`inUse` is mainted in order to ensure that `maxConnections` represents a
strict upper bound. Since we can not perform I/O operations within STM
transactions, we increment `inUse` within STM and after performing the
I/O operation which creates a new resource, the resource is put into the
sequence in a separate transaction. In between these two transactions a
concurrently executing one may see the incremented `inUse` value but not
the correct number of resources. In this case we should `retry` until we
eventually see the invariant restored.

To ensure that for every `inUse` increment there is either a new
additional resource or a `inUse` decrement in case of errors, we run
`take1` within `uninterruptibleMask`, only unmasking the actual resource
creation which `onException` decrements `inUse`.