Update to retry-0.7
Upstream has merged the PR which provides retry policies with access to
the retry counter (as well as other things), so the internal fork of the
retry package can go away.  It also allows/requires the retry policies
to run in the user's monad (it provides a higher-rank synonym to operate
in an unspecified monad to ease the transition, but trying to use it ran
into the unsupported land of impredicative types), so `RetrySettings`
now holds a `RetryPolicyM Client`.  That in turn means `Settings.hs`
needs to know that `Client` exists and is a monad, though nothing else.